% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{dynamic_knapsack}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{, is the input weight}

\item{W}{, is the input value}
}
\value{
list , returns a list that contain the maximum value and the list of elements chosen
}
\description{
dynamic_knapsack function to solve the knapsack problem
}
