% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6R.R
\docType{package}
\name{Lab6R}
\alias{Lab6R}
\alias{Lab6R-package}
\title{Lab6R: This is the package for R lab 6
This package contain 3 function:}
\description{
Lab6R: This is the package for R lab 6
This package contain 3 function:
}
\section{brute_force_knapsack function}{

Use brute force approach to solve the knapsack problem.
The result is a list that contain the maximum value and the list of elements chosen
}

\section{dynamic_knapsack function}{

The same with the previous function, but use dynamic approach to solve the problem
}

\section{greedy_knapsack fucntion}{

The same with the first function, but use greedy approach to solve the problem
}

