% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{, is the input weight}

\item{W}{, is the input value}
}
\value{
list , returns a list that contain the maximum value and the list of elements chosen
}
\description{
brute_force_knapsack function to solve the knapsack problem
}
